Finished consolidating orderbook packages.
