+++ src/performance/scala-2.11/markets/orderbooks/concurrent/OrderBookMicroBenchmark.scala
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.mutable.orderbooks
+package markets.orderbooks.concurrent
 
 import markets.tradables.Security
 import markets.tradables.orders.ask.{AskOrder, LimitAskOrder}
@@ -23,8 +23,8 @@ import org.scalameter.{Bench, Gen}
 import scala.util.Random
 
 
-/** Performance tests for the `SortedOrderBook` class. */
-object SortedOrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
+/** Performance tests for the `ConcurrentOrderBook` class. */
+object OrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
 
   import markets.RandomOrderGenerator._
 
@@ -36,7 +36,7 @@ object SortedOrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
 
   /** Generates a collection of `ConcurrentOrderBook` instances of increasing size. */
   val orderBooks = for { size <- sizes } yield {
-    val orderBook = SortedOrderBook[AskOrder](validTradable)
+    val orderBook = OrderBook[AskOrder](validTradable)
     val orders = for (i <- 1 to size) yield randomAskOrder(prng, tradable = validTradable)
     orders.foreach( order => orderBook.add(order) )
     orderBook
@@ -49,7 +49,7 @@ object SortedOrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
     exec.jvmflags -> List("-Xmx2G")
     ) in {
 
-    /** Adding an `Order` to a `SortedOrderBook` should be an `O(log n)` operation. */
+    /** Adding an `Order` to an `OrderBook` should be an `O(1)` operation. */
     measure method "add" in {
       using(orderBooks) in {
         orderBook =>
@@ -65,26 +65,19 @@ object SortedOrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
       }
     }
 
-    /** Finding an `Order` in a `SortedOrderBook` should be an `O(n)` operation. */
+    /** Finding an `Order` in an `OrderBook` should be an `O(n)` operation. */
     measure method "find" in {
       using(orderBooks) in {
         orderBook => orderBook.find(order => order.isInstanceOf[LimitAskOrder])
       }
     }
 
-    /** Removing an `Order` from a `SortedOrderBook` should be an `O(log n)` operation. */
-    measure method "remove(order)" in {
+    /** Removing an `Order` from an `OrderBook` should be an `O(1)` operation. */
+    measure method "remove" in {
       using(orderBooks) in {
         orderBook =>
           val (uuid, _) = orderBook.existingOrders.head
           orderBook.remove(uuid)
-      }
-    }
-
-    /** Removing the head `Order` from a `SortedOrderBook` should be an `O(log n)` operation. */
-    measure method "remove()" in {
-      using(orderBooks) in {
-        orderBook => orderBook.remove()
       }
     }
