+++ src/performance/scala-2.11/markets/orderbooks/mutable/OrderBookMicroBenchmark.scala
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.mutable.orderbooks
+package markets.orderbooks.mutable
 
 import markets.tradables.Security
 import markets.tradables.orders.ask.{AskOrder, LimitAskOrder}
