+++ src/performance/scala-2.11/markets/orderbooks/mutable/PriorityOrderBookMicroBenchmark.scala
@@ -13,8 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.parallel.mutable.orderbooks
+package markets.orderbooks.mutable
 
+import markets.RandomOrderGenerator
 import markets.tradables.Security
 import markets.tradables.orders.ask.{AskOrder, LimitAskOrder}
 import org.scalameter.api._
@@ -23,37 +24,37 @@ import org.scalameter.{Bench, Gen}
 import scala.util.Random
 
 
-/** Performance tests for the `OrderBook` class. */
-object OrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
+/** Performance tests for the `PriorityOrderBook` class. */
+object PriorityOrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
 
-  import markets.RandomOrderGenerator._
+  import RandomOrderGenerator._
 
   val prng = new Random(42)
 
-  val validTradable = Security(uuid())
+  val tradable = Security(uuid())
 
   val sizes = Gen.exponential("Number of existing orders")(factor=10, until=1000000, from=10)
 
-  /** Generates a collection of OrderBooks of increasing size. */
+  /** Generates a collection of SortedOrderBooks of increasing size. */
   val orderBooks = for { size <- sizes } yield {
-    val orderBook = OrderBook[AskOrder](validTradable)
-    val orders = for (i <- 1 to size) yield randomAskOrder(prng, tradable = validTradable)
+    val orderBook = PriorityOrderBook[AskOrder](tradable)
+    val orders = for (i <- 1 to size) yield randomAskOrder(prng, tradable = tradable)
     orders.foreach( order => orderBook.add(order) )
     orderBook
   }
 
-  performance of "OrderBook" config (
-    reports.resultDir -> "target/benchmarks/markets/engines/orderbooks/OrderBook",
+  performance of "PriorityOrderBook" config (
+    reports.resultDir -> "target/benchmarks/markets/engines/orderbooks/PriorityOrderBook",
     exec.benchRuns -> 200,
     exec.independentSamples -> 20,
     exec.jvmflags -> List("-Xmx2G")
     ) in {
 
-    /** Adding an `Order` to an `OrderBook` should be an `O(1)` operation. */
+    /** Adding an `Order` to a `PriorityOrderBook` should be an `O(log n)` operation. */
     measure method "add" in {
       using(orderBooks) in {
         orderBook =>
-          val newOrder = randomAskOrder(prng, tradable=validTradable)
+          val newOrder = randomAskOrder(prng, tradable=tradable)
           orderBook.add(newOrder)
       }
     }
@@ -65,14 +66,14 @@ object OrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
       }
     }
 
-    /** Finding an `Order` in an `OrderBook` should be an `O(n)` operation. */
+    /** Finding an `Order` in an `PriorityOrderBook` should be an `O(n)` operation. */
     measure method "find" in {
       using(orderBooks) in {
         orderBook => orderBook.find(order => order.isInstanceOf[LimitAskOrder])
       }
     }
 
-    /** Removing an `Order` from an `OrderBook` should be an `O(1)` operation. */
+    /** Removing an `Order` from a `PriorityOrderBook` should be an `O(log n)` operation. */
     measure method "remove(order)" in {
       using(orderBooks) in {
         orderBook =>
@@ -81,7 +82,7 @@ object OrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
       }
     }
 
-    /** Removing the head `Order` from an `OrderBook` should be an `O(1)` operation. */
+    /** Removing the priority `Order` from a `PriorityOrderBook` should be an `O(log n)` operation. */
     measure method "remove()" in {
       using(orderBooks) in {
         orderBook => orderBook.remove()
