+++ src/performance/scala-2.11/markets/orderbooks/parallel/mutable/OrderBookMicroBenchmark.scala
@@ -13,18 +13,17 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.parallel.concurrent.orderbooks
+package markets.orderbooks.parallel.mutable
 
-import markets.tradables.orders.ask.{AskOrder, LimitAskOrder}
 import markets.tradables.Security
-
+import markets.tradables.orders.ask.{AskOrder, LimitAskOrder}
 import org.scalameter.api._
 import org.scalameter.{Bench, Gen}
 
 import scala.util.Random
 
 
-/** Performance tests for the `ConcurrentOrderBook` class. */
+/** Performance tests for the `OrderBook` class. */
 object OrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
 
   import markets.RandomOrderGenerator._
@@ -35,7 +34,7 @@ object OrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
 
   val sizes = Gen.exponential("Number of existing orders")(factor=10, until=1000000, from=10)
 
-  /** Generates a collection of `ConcurrentOrderBook` instances of increasing size. */
+  /** Generates a collection of OrderBooks of increasing size. */
   val orderBooks = for { size <- sizes } yield {
     val orderBook = OrderBook[AskOrder](validTradable)
     val orders = for (i <- 1 to size) yield randomAskOrder(prng, tradable = validTradable)
@@ -74,11 +73,18 @@ object OrderBookMicroBenchmark extends Bench.OnlineRegressionReport {
     }
 
     /** Removing an `Order` from an `OrderBook` should be an `O(1)` operation. */
-    measure method "remove" in {
+    measure method "remove(order)" in {
       using(orderBooks) in {
         orderBook =>
           val (uuid, _) = orderBook.existingOrders.head
           orderBook.remove(uuid)
+      }
+    }
+
+    /** Removing the head `Order` from an `OrderBook` should be an `O(1)` operation. */
+    measure method "remove()" in {
+      using(orderBooks) in {
+        orderBook => orderBook.remove()
       }
     }
